Improvements to search
