+++ amplify/backend/function/tmhLogLivestreams/custom-policies.json
@@ -1,0 +1,6 @@
+[
+  {
+    "Action": [],
+    "Resource": []
+  }
+]
