+++ amplify/backend/function/tmhLogLivestreams/function-parameters.json
@@ -1,0 +1,3 @@
+{
+  "lambdaLayers": []
+}
