+++ amplify/backend/function/tmhLogLivestreams/src/event.json
@@ -1,0 +1,5 @@
+{
+  "key1": "value1",
+  "key2": "value2",
+  "key3": "value3"
+}
