+++ amplify/backend/function/tmhLogLivestreams/src/index.js
@@ -1,0 +1,12 @@
+/**
+ * @type {import('@types/aws-lambda').APIGatewayProxyHandler}
+ */
+exports.handler = (event) => {
+  console.log(`EVENT: ${JSON.stringify(event)}`);
+  for (const record of event.Records) {
+    console.log(record.eventID);
+    console.log(record.eventName);
+    console.log('DynamoDB Record: %j', record.dynamodb);
+  }
+  return Promise.resolve('Successfully processed DynamoDB record');
+};
