+++ amplify/backend/function/tmhLogLivestreams/src/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "tmhloglivestreams",
+  "version": "2.0.0",
+  "description": "Lambda function generated by Amplify",
+  "main": "index.js",
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "@types/aws-lambda": "^8.10.92"
+  }
+}
