+++ src/components/RenderRouter/SearchPage/SearchItemPill.tsx
@@ -1,0 +1,4 @@
+import './SearchItem.scss';
+export default function Pill({ title }: { title: string }) {
+  return <span className="SearchItemPill">{title}</span>;
+}
